Changes commit

+++ src/main/java/ru/netology/javaqa/hwarrays/services/StatsService.java
@@ -1,7 +1,7 @@
 package ru.netology.javaqa.hwarrays.services;
 
 public class StatsService {
-    public int minSales(int[] sales) { // номер месяца с минимальными продажами среди просмотренных ранее
+    public int minSales(long[] sales) { // номер месяца с минимальными продажами среди просмотренных ранее
         int minMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] <= sales[minMonth]) { // значит, в рассматриваемом i-м месяце продаж меньше
@@ -11,7 +11,7 @@ public class StatsService {
         return minMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
 
-    public int maxSales(int[] sales) { // номер месяца с максимальными продажами среди просмотренных ранее
+    public int maxSales(long[] sales) { // номер месяца с максимальными продажами среди просмотренных ранее
         int maxMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] >= sales[maxMonth]) { // значит, в рассматриваемом i-м месяце продаж больше
@@ -21,25 +21,25 @@ public class StatsService {
         return maxMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
 
-    public int sumSales(int[] sales) { // Сумма всех продаж
-        int sum = 0;
-        for (int t : sales) { // цикл для обхода каждого элемента массива
+    public long sumSales(long[] sales) { // Сумма всех продаж
+        long sum = 0;
+        for (long t : sales) { // цикл для обхода каждого элемента массива
             sum = sum + t; //суммирование каждого элемента массива
         }
         return sum;
     }
 
-    public int averageSales(int[] sales) { // Средняя сумма продаж в месяц
-        int sum = 0;
-        int average = 0;
-        for (int t : sales) { // цикл для обхода каждого элемента массива
+    public long averageSales(long[] sales) { // Средняя сумма продаж в месяц
+        long sum = 0;
+        long average = 0;
+        for (long t : sales) { // цикл для обхода каждого элемента массива
             sum = sum + t;
             average = sum / 12;
         }
         return average;
     }
 
-    public int averageSalesMin(int[] sales) { // Количество месяцев с продажами ниже среднего (меньше 15)
+    public int averageSalesMin(long[] sales) { // Количество месяцев с продажами ниже среднего (меньше 15)
         int count = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] < 15) {
@@ -49,7 +49,7 @@ public class StatsService {
         return count;
     }
 
-    public int averageSalesMax(int[] sales) { // Количество месяцев с продажами выше среднего (меньше 15)
+    public int averageSalesMax(long[] sales) { // Количество месяцев с продажами выше среднего (меньше 15)
         int count = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] > 15) {
